news-backend/requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
firebase-admin
pydantic
